add missing libs 2

+++ controls/VTDesign/Text.cpp
@@ -1056,7 +1056,7 @@ LRESULT CVTText::DefWindowProc(UINT message, WPARAM wParam, LPARAM lParam)
 {
 	if( message == WM_GETTEXT )
 	{
-		lstrcpyn((LPTSTR)lParam, (LPCTSTR)m_strText, wParam);
+		lstrcpyn((LPTSTR)lParam, (LPCTSTR)m_strText, (int)wParam);
 		if ((int)wParam > m_strText.GetLength())
 		wParam = m_strText.GetLength();
 		return TRUE;
